rpa-agent/requirements.txt
pywinauto>=0.6.8
Pillow>=10.0.0
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
pytest>=8.0.0
